clear code, rename js file, hide or display edit and delete btn if admin/user
