+++ client/index.html
@@ -117,8 +117,7 @@
 	</div>
 
 	<!-- SCRIPT JS -->
-	<script type="module" src="./js/modal.js"></script>
+	<script type="module" src="./js/modal-display.js"></script>
 	<script type="module" src="./js/app.js"></script>
-	<script type="module" src="./js/processingModal.js"></script>
 </body>
 </html>
