+++ client/js/modal-display.js
@@ -1,5 +1,13 @@
 const modal = document.getElementById('new-module-form-container')
 
+// display modal
+export const setup = () => {
+  if (document.isAdmin) {
+    const plusBlock = document.getElementById('block-plus')
+    plusBlock.addEventListener('click', displayModal)
+  }
+}
+
 export const displayModal = () => {
   modal.style.display = 'block'
 }
