+++ mocks/blocks/blocks.json
@@ -43,50 +43,5 @@
     "color": "#ffffe6",
     "titleColor": "#292e2a",
     "position": 3
-  },
-  {
-    "id": 6,
-    "title": "COCOCO",
-    "url": "http://test.com",
-    "icon": "img/icon-livecodings.png",
-    "color": "#FFFFE6",
-    "titleColor": "#292e2a",
-    "position": 6
-  },
-  {
-    "id": 7,
-    "title": "COCOCO",
-    "url": "http://test.com",
-    "icon": "img/icon-livecodings.png",
-    "color": "#F3F3F3",
-    "titleColor": "#292e2a",
-    "position": 7
-  },
-  {
-    "id": 8,
-    "title": "ezfze",
-    "url": "http://test.com",
-    "icon": "img/icon-dojos.png",
-    "color": "#FFFF57",
-    "titleColor": "#292e2a",
-    "position": 8
-  },
-  {
-    "id": 9,
-    "title": "COCOCO",
-    "url": "http://test.com",
-    "icon": "img/icon-groups-blanc.png",
-    "color": "#343936",
-    "titleColor": "white",
-    "position": 9
-  },
-  {
-    "id": 10,
-    "title": "zefzef",
-    "url": "http://test.com",
-    "icon": "img/icon-livecodings-blanc.png",
-    "color": "#343936",
-    "titleColor": "white",
-    "position": 10
   }
 ]
